Attempted to merge
Removed a ton of redundant code that made reading through
the player controller difficult, and future changes
even more difficult.
New player controller prefab is in folder "Player".
Scanner mesh is now a cone, and becomes visible when scanning
is in progress as feedback for the player.
Mouse cursor locks on game start to the droid. Press Escape to unlock it.