Vague AI or something.
"Mushroom" thing sits in a corner and shoots a sphere representing a range of spores.
On contact, player is covered with spores broadcasting it as threat.
Seeker mockup chases player when they're both near and have threat spores.
Fancy particle effects when the player is covered in spores.
Seriously. That probe controller is just dysfunctional. Super slow in Unity, too fast in a build. Probably needs a deltaTime somewhere.